aeronet/requirements.txt
# Core deep learning
torch>=2.1
lightning>=2.2

# Data + IO
numpy>=1.24
pyvista>=0.44
pyyaml>=6.0
tqdm>=4.65

# Optional: only for the test scripts
pytest>=7.0
